Precompute hash and lastmod during the GenerateJavascript stage.
This is useful when running from jars (i.e. no source present) for two reasons.
First, it's just plain faster, instead of scanning the classpath and testing
all ~10^4 entries' names for a string match.  Second, it's actually correct,
whereas the classpath scan is buggy because (a) right now, we're missing things
and end up with an empty set (/aura/resources/** vs. aura.resources.**, etc.),
and (b) it's foolable, because a non-aura classpath thing might declare items
in our space, but which were not part of the compile set.

It's neither more nor less correct against spoofing, though: one could mutate
either the fwuid properties file OR, today, the source, and create a mismatch.

@rev goliver@
@bug ???@