Add replace() to AuraHistoryService.
This is needed to fix multiple issues where back actions display incorrect or blank pages.
Since this is an addition of a new function, it shouldn't destabilize any existing code.
For clients which do not support replaceState or location.replace, we fallback to setting
the hash explicitly, the current behavior of set.